[Refactor/#180] add valid annotation check time form
