+++ src/main/java/com/example/waggle/web/dto/pet/PetRequest.java
@@ -1,7 +1,7 @@
 package com.example.waggle.web.dto.pet;
 
 import com.example.waggle.domain.member.entity.Gender;
-import com.example.waggle.global.annotation.valid.ValidEnum;
+import com.example.waggle.global.annotation.valid.enums.ValidEnum;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
