+++ src/main/java/com/example/waggle/web/dto/siren/SirenRequest.java
@@ -2,7 +2,7 @@ package com.example.waggle.web.dto.siren;
 
 import com.example.waggle.domain.board.siren.entity.SirenCategory;
 import com.example.waggle.domain.member.entity.Gender;
-import com.example.waggle.global.annotation.valid.ValidEnum;
+import com.example.waggle.global.annotation.valid.enums.ValidEnum;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
